src/analyzer/requirements.txt
requests==2.31.0
paho-mqtt==1.6.1
influxdb-client>=1.38.0
scikit-learn>=1.3.0
numpy>=1.24.0